requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.4
django-cors-headers==4.2.0
django-filter==23.2
django-livereload-server==0.4
djangorestframework==3.14.0
Markdown==3.4.4
numpy==1.25.2
pandas==2.0.3
Pillow==10.0.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tornado==6.3.2
tzdata==2023.3
